Remove special casing of type variables: getAnnotations/isAnnotated now only work on the type variable itself. New methods to get the "effective" type were added.
When determining the supertypes of type variables and wildcards create
a deep copy, as some other method modifies the types in the supertype
list.

Descend into arrays when replacing types.

Ensure that the component type of arrays is correctly output:
use the simple class name and output annotations on bounds by using
the AnnotatedTypeMirror instead of the underlying type.
Does this create a problem somewhere I didn't see?

Unify coding style.