Don't build CFG for abstract methods.
Skip over ANNOTATION_TYPE members of a class.

Fix error in narrowing for assignment conversion.  Narrowing should
only be allowed for byte, char, short and their boxed versions.

Add ThrowNode for throw statements.

Change representation of exception from Class<? extends Throwable>
to TypeMirror because we need to represent exceptions defined in
the program being analyzed which may not be present in the checker itself.