Close the lub branch.
Here is Stefan's explanation:

At some point we had problems that taking the LUB of two annotated types
was not computed correctly, because we initially took the somewhat naive
approach of taking the LUB of the underlying Java type and just adding the
LUB of the annotations on both types.

To fix this problem, we tried to implement the same LUB strategy as javac
does for Java types in Types and have the same computation that also
considers annotations in AnnotatedTypes.  Unfortunately, this turned out to
be rather challenging because the computation is not straight-forward, and
the existing code hardly documented.  This was done in the lub branch.

In the end we abandoned this approach (we never got it working) and decided
to fix the our existing LUB computation.  Now, we consider all necessary
cases and also add the right annotations to element types, bounds, etc.
This is important for taking the LUB of array types, type variables,
wildcards and so on.  This required some special casing and trial-and-error
to get all the behavior we want, but in the end it seemed easier than
trying to understand and reimplement the javac LUB method.  It also seems
to work reliably now.