fix errors about incompatible type variables by defining proper hashCode/equals for QualifiedTypeMirror
Previously (when QTM.hashCode was using the default implementation), the same
QTM would be given a new typeIndex for every call to TypeMirrorConverter's
getAnnotatedType or bindTypes methods.  This caused problems with type
variables, because converting a type variable ATM to and from a QTM would
produce a distinct non-.equals() ATM on each call.