Resolved 3 todos in the StubParser: 1. TODO: Should throw exception?! No, we now use the stubWarning method.
2. TODO: instead of failing, can we try merging the information from multiple stub files?
No, this is a generic method that adds entries to a map that is passed as an arugement,
so there is no sensable way to merge.  Also, stub file merge is implemented else where.

3.  TODO: The only compile time constants supported here are Strings
This method also supports boolean constant now.  I changed the todo to a comment and
add this limiation to the manual.