Fixes to make the loader work with loading from any classpath
The loader will now scan all available classpaths to see if any of the jars or directories contain the qual package for a checker. If so, will will mark that jar or directory as the source for loading qual directory classes. If no jar or directory contains the qual package (which can happen for framework checkers) then it will simply not load anything.