Updated the loader for some speed improvements
Updated loader to first check and see if the classloader is available, if so it uses the old logic which will directly retrieve the resource URL of the qual package. If the classloader is unavailable (ie the bootstrap classloader was used to load checker) then it uses the new logic to scan classpaths. These changes are in constructor method.

Also there's code formatting cleanup throughout the rest of the file.