fixes for jenkins
Fixed an environment variable definition bug in annotationclassloader/Makefile for the location of javac.jar
It was previously assumed that the jsr308-langtools project was located in the same directory as the checker-framework project, and it had used relative paths to point to javac.jar.
It will now receive the proper jsr308-langtools project directory from ant if called by ant script.
If it is directly executed from command line, it will still utilize the old directory (users are asked to update the makefile in such a scenario to reflect their project configurations)

Also updated readme, makefile, and build target so that the ordering is always the loading from directory test before the loading from jar test