No default qualifier stack trace message enhancement
In the scenarios where @DefaultQualifierInHierarchy or @DefaultFor(DefaultLocation.OTHERWISE) are not found in any of the qualifiers added to the supported qualifier set, or if that set is empty, the stack trace does not display how many or which qualifiers are examined. This fix updates the stack trace output to print out the number of qualifiers in the set as well as the canonical names of the qualifiers.