Don't try to infer type arguments for raw types
Because the assignment context is sometimes null during dataflow, type agrument inference for rawtypes fails and is cached. So raw types behave differently if caching is turned off.  For now do not infer type arguments for raw types.  Doing this leads to many errors, so the ignoreRawTypeArguments option is always used.  Type argument inference for raw types should be improved and this option should always be useed.  (See issue #635)