Adds whole-program inference
With this feature enabled, method return types and field types are stored in .jaif
files. The types in the .jaif files get updated at return statements and
assignments. Note that this works for both private and non-private methods and
fields.

After generating the .jaif files, the insert-annotations-to-source tool can be
used to add annotations to .java files. I created a script that iteratively runs
inference and adds annotations from .jaif files to the .java files on the
program until the .jaif files from one execution are similar to the .jaif files
from the previous execution.