Implement refinements based on binary comparison operators in Value Checker
And also create ArrayLen annotations for arrays when their lengths have more precise IntVal annotations.

Fixes kelloggm#131

Note that the only thing that changed in ValueAnnotatedTypeFactory#resultAnnotationHandler is that I moved it from the TreeAnnotator to the ATF itself. I did that so that I could write ValueAnnotatedTypeFactory#createAnnotationFromResults, which wraps resultAnnotationHandler so that ValueTransfer can use it.