Issue more warnings if explicitly written Value Checker annotations are not wellformed (#1221)
Issue a warning if an explicit array len has a negative value.  Issue a warning if any Value annotation uses an empty array.  I also refactored the code that issue warnings on explicit annotations to use AnnotationMirror so that the code is easier to read.

Also, annotations that are not wellformed are converted to Top, expect for @ArrayLen with negative numbers is converted to bottom.