Added some missing imports
While I appreciate the brevity of the opening example, it is very
confusing to show Java source code to a user without import lines for
the relevant classes.  Finding specific classes in the sea of JavaDocs
for any large project can be quite tedious, and we may as well save
people that tedium.