Ensure that `make` succeeds even when ../../checker/api is missing
Previously, if ../../checker/api had not been built, the rule to build
../api would create a link to a non-existent folder.  This creates
problems on rebuild: Make sees links to non-existent targets as
non-existent and always tries to rebuild them, but `ln` will fail if
the link already exists (even if the target does not).

This commit works around the issue by adding the "-f" flag to `ln` so
that the link is quietly rebuilt each time.