Create MinLen annotations from Positive annotations (#1259)
This solves kelloggm#66 and similar bugs by inferring `@MinLen` annotations from arrays created using a `@Positive` int.

This creates a new type in the Value Checker called `@IntRangeFromPositive`. It is only introduced by when an `@Positive` annotation is read (it is aliased), and it is treated by dataflow/lub/glb/isSubtype/etc. as if it were an `@IntRange(from = 1)` annotation. `IntRange`s derived from it are normal `@IntRange` annotations. However, I also changed the common assignment checks so that if the left side has type `@IntRangeFromPositive`, the standard common assignment checks do not occur (in ValueVisitor). Writing `@IntRangeFromPositive` in code directly is an error. `@IntRangeFromPositive` is also not written into bytecode.

I also added a true bottom type to the LBC, because the Value Checker was trying to add `@IntRangeFromPositive` annotations to null.