Refine minimum string length after startsWith and endsWith (#1432)
This is a part of the Bounded-size strings GSoC project. 

With these changes, the minimum length of the receiver string in the then-branch is refined to the minimum length known for the argument (based on a stored annotation or the length of a string literal).

The purpose is to support code like:

if(str.startsWith("x")){
 String after = str.substring(1);
}

This pattern appears often in plume-lib and causes warnings from the Index Checker with support for strings. With the added refinement, the checker will see that the string is at least as long as the prefix or suffix and so the index is valid for the string.