checker-qual.jar now uses Java 8 byte code (#1537)
`checker-qual.jar` used to use Java 7 byte code.  The utility classes, such as `NullnessUtil`, use annotations in comments so those annotations appear in the byte code, but are ignored by the standard Java 8 compiler.  So, `checker-qual.jar` uses Java 8 byte code so that it can be used with the standard Java 8 compiler.  This fixes #1510. 

Additionally, this adds a new jar, `checker-qual7.jar`, that is exactly the same as `checker-qual.jar`, but uses Java 7 byte code.  The annotations in the utility classes remain in  comments to that this jar can be built.  They are removed from comments when building `checker-qual-sources.jar`.