Fix SameLen LUB (#1559)
SameLen's LUB was incorrect. At a join point, it returned an annotation containing all the arrays in both input annotations. The correct LUB only includes arrays in both. For instance,

void test(int[] a, boolean cond) {
  int[] b;                // a is @SameLen("a") now
  if (cond) b = a;  // inside this if block, a is @SameLen("a", "b")
  int[] c = a;         // after this join point, a should be @SameLen("a"). 
}                           // Using the old LUB, a is @SameLen("a", "b")
This change fixes this mistake. I also removed a helper method that exists in the standard library, and called the standard library instead. This also fixes kelloggm#162