Remove org.checkerframework.framework.qual.Dependent (#1767)
Only the Nullness Checker supported this annotation.  It's not well documented or well implemented.  (One of the two tests using this annotation is skipped.)

It is implemented using the `GeneralAnnotatedTypeFactory` which supports storing
annotations that are not qualifiers in AnnotatedTypeMirrors.  It's causing the test failures in #1766.  I moved it `org.checkerframework.common.util.debug.TypeOutputtingChecker` because `TypeOutputtingChecker` still uses it.