Convert the build scripts from Ant to Gradle. (#1871)
I tried to make the Gradle tasks do the same thing as the Ant tasks did.  However, since we are ending support for JSR308, I removed all references to it in the build scripts except for the jtreg tests, which is explained in a comment in that task.  CheckerMain still uses it though because plume-lib still uses annotations in comments and because we haven't removed mentions of annotations in comments from the manual.

See [developer-manual.html](https://github.com/typetools/checker-framework/pull/1871/files#diff-521ff72506c5df28427c8b58da6cb293) for overview of the Gradle tasks.

The release scripts need to be updated for this change, but that will happen after this is merged.

After this is merged, we will remove the ant build scripts.  After that, we should relocate source and tests files so that each source set is in its own directory.