Use ArrayRange and ArrayLen in LessThan Checker. (#2003)
Fixes  https://github.com/kelloggm/checker-framework/issues/225.

The test case was failing because bar.length isn't in the Value Checker's store until after 0. So the Value Checker doesn't have a value for bar.length. But is does have an ArrayRange for bar, so I changed the code to use that.