Reflection Resolution: (#2112)
Verify that the first argument to an invoke call is a subtype of the possible class.

The fix for #2107 uncovered this bug. The test cases with this were crashing that fix. This fixes the crash. With the crash fixed, the test case with this correctly issued errors. But, since the point of those tests were to check the reflection resolution was improving the precision of the checker, I had to change them to something that would pass.