Rename two methods
isPolymorphicQualified => hasPolymorphicQualifier
Previously, the code had isPolymorphicQualified and getPolymorphicQualifier.
I feel that parallelism between hasPolymorphicQualifier and getPolymorphicQualifier is nicer.

getPolymorphicQualifierElement => getPolymorphicQualifierTop
The new name better reflects the actual semantics.
We could instead require that the argument to @PolymorphicQualifier must be exactly
top, and in that case this renaming would not be necessary.