Explicit annotations on type parameter upper bounds. (#2501)
The changes in #2472 caused daikon-type-check to fail.
Here's the error:

daikon/diff/Node.java:22: error: [assignment.type.incompatible] incompatible types in assignment.
    this.userObject = userObject;
                      ^
  found   : @NonNull @NonRaw Pair<CONTENT[ extends @NonNull @NonRaw Object super @NonNull @NonRaw Void], CONTENT[ extends @NonNull @NonRaw Object super @NonNull @NonRaw Void]>
  required: @NonNull @NonRaw Pair<CONTENT[ extends @NonRaw @Nullable Object super @NonNull @NonRaw Void], CONTENT[ extends @NonRaw @Nullable Object super @NonNull @NonRaw Void]>
Here is the relevant snippet of the Node.java file:

public abstract class Node<CONTENT extends @Nullable Object, CHILD> {

  private List<CHILD> children = new ArrayList<CHILD>();
  // Nonsensical for RootNode
  private Pair<CONTENT, CONTENT> userObject;

  public Node(Pair<CONTENT, CONTENT> userObject) {
    this.userObject = userObject;
  }
The error only seems to happen when the tree for the element isn't available, so I'm not sure how to create a test case for this. However, I have verified locally that this pull request fixes the bug.