Postconditions should refine the refined type of an expression (#2628)
Postconditions should refine the refined type of an expression so that the type of an expression can become more and more precise. This is done by storing the glb of the precondition qualifier and the value in the store. If the glb is bottom, just use the new value.

Fixes #2619.