Adds @NoQualifierParameter and @HasQualifierParameter on packages (#19)
This adds the ability to turn on @HasQualifierParameter for all classes in a package by writing @HasQualifierParameter as a package annotation so that it doesn't need to be written on each class individually.

Also adds a new annotation, @NoQualifierParameter, which indicates no implicit qualifier parameter. The purpose is so that if @HasQualifierParameter is enabled by default for a package, it can be disabled for specific classes.