[dataflow] Class literal nodes should have exceptional edges
Summary:
Absence of exceptional edges leads to missing CFG nodes in a code like
this:
```
try {
  var cls = MyClass.class;
  ...
} catch (NoClassDefFoundError e) {
  ...
}
```

According to JLS evaluating a class literal can fail in a number of
ways which are handled by this patch.

Test plan:
Added a test case under nullness to check if the types inside `catch`
get refined.