Remove `useLocalScope` when parsing JavaExpressions. (#4276)
JavaExpressionParseUtil#parse no longer takes the boolean useLocalScope. Instead, if the JavaExpression should be parsed as if it were written at a given TreePath, then a non-null localVarPath is passed. If localVarPath is not needed, then no path should be passed. (Internally, the JavaExpression parser uses the path to the current compilation unit when calling javac APIs.). 

Also, added viewpointAdaptQualifierFromContract to DependTypesHelper.