Changed BaseTypeVisitor to set the language level for parsing to match the current source level of the compilation.
This seems wise generally, and will become necessary for Java 16 support to function correctly, as otherwise Java 16 features won't be accepted by the parser.

The other parser uses in the source code currently use Java 8 only.  This may also need to change (what if you want a stub involving a record?) but I have not changed those in this commit/PR.