Updating inference.tex (wpi-many.sh docs)
The format of the GitHub repository URL and git hash pairs provided as input
to `wpi-many.sh` is not clear. The current documentation, i.e.,

> Use wpi-many.sh to run whole-program inference on every Ant, Gradle, or Maven project in a list of (GitHub repository URL, git hash) pairs.

Available [here](https://checkerframework.org/manual/#wpi-many) suggests that
the the elements of the pairs should be comma-separated, when it in fact should
be whitespace-separated.