Improve exception handling in `CalledMethodsTransfer.accumulate`
The call to `s.insertValue` may silently do nothing if the new
annotation is not a super- or sub-type of the existing one.  This can
actually happen in some cases.

For example, in the new test case, at the call to `r.b()`:

 - The incoming transfer input lists {"a"} as the called methods.
 - The result of the transfer function for the regular path produces
   only {"b"} as the called methods (since the call to `b()` may
   modify `r`, invalidating the set of called methods).
 - The exceptional store handling begins with every exceptional store
   equal to the input store. When it tries to insert {"b"} over the
   input value {"a"}, nothing happens since neither value is more
   specific.

This change makes the accumulation completely overwrite the exceptional
output with the regular output instead of allowing a silent no-op.