At (pseudo) assignments, use the existing strength of any obligations
In PR #6241 (commit 5b5c9e53) I was very cautious about the value of
the `whenToEnforce` field on obligations.  It is always safe to use
`MethodExitKind.ALL`, and so in most cases that's what I used.

However, having had some time to code with the new change, it's clear
that `MethodExitKind.ALL` is too strong for (pseudo) assignments.  The
left hand side should have the same `whenToEnforce` as the right-hand
side.

See the new test case `MustCallAliasInitializeWithOwningParameter`.
The pattern there is (in my experience) relatively common, but it
does not verify with the latest Checker Framework release.

This change alters the behavior of a few of the `InputOutputStreams`
test cases, since those tests take `@Owning` arguments instead of
allocating new objects.  I have added copies of the changed tests to
retain their intended behavior; the copies allocate sockets instead of
using `@Owning` parameters.