Skip `TreeUtils.toStringTruncated` when debugging is disabled
The `getAnnotatedType` method is one of the most-called in the Checker
Framework.  It spends a lot of time and allocates a lot of bytes
calling `TreeUtils.toStringTruncated`, but the results of those calls
are only used when an internal debug flag is enabled (uncommon).

To fix the problem, this commit makes the call to `toStringTruncated`
automatic in `logGat`, so the call sites do not need to do the string
conversion and truncation themselves.

This optimization opportunity was discovered by profiling on a small
project (50-100 source files).  For that workload,

 - `toStringTruncated` contributed roughly 5% of total run time, and
   most of those calls came from `getAnnotatedType`.
 - `toStringTruncated` contributed roughly 25% (!!!) of total
   allocated bytes.

I don't yet know if this change will have such a big impact on other
projects, but it seems like a straightforward improvement.