Avoiding crash when Non-Empty Checker is run
The Non-Empty Checker admits a compiler option `-ArunAsOptionalChecker`, which
directs it to run as an Optional Checker with increased precision wrt.
container types.

A boolean variable in `NonEmptyChecker` was set to whether this option was
passed or not. However, it was used *before* it was set (under the assumption
that `getImmediateSubcheckerClasses` was executed *after* `initChecker()`,
which was not the case.

As a result, the assertion in `NonEmptyChecker#getMethodsToVerify` failed and
crashed the checker.

This change avoids using a variable to store the value of the key, and instead
reads the value of the key directly.