Re-naming the (old) `OptionalChecker` to `OptionalWithoutNonEmptyChecker`
This change is required in preparation for re-naming the current
`NonEmptyChecker` (which behaves like the Optional Checker with greater
precision when run in the `-ArunAsOptionalChecker` modality) to the
`OptionalChecker`, i.e., for it to become the new, default Optional Checker.

Note that there are also changes to the keys used to suppress warnings, e.g.,

  `optional:parameter` becomes `optionalwithoutnonempty:parameter`

to avoid breaking the build; these suppression keys should be switched back
to the original key after the re-naming is complete.