All primitive types are `@Interned`
Fixes #6371.

The interning checker does not apply `@Interned` to every occurrence of
every primitive type.  This allows some language constructs to
fabricate non-interned primitive values, leading to false positives.

For instance, in the new `loopVariables` test, the loop variable `i`
was assigned the type `@UnknownInterned long`.  Thus `i` could not be
used as an argument to `addExact` which requires interned arguments.