Fix resource leak unsoundness due to `finally { throw }`
This commit fixes a subtle unsoundness in the resource leak checker
concerning try-finally blocks where the `finally` block might throw an
exception.

The simplest variant of this confusing control flow is shown in the
first test case in FinallyThrowsException.java:

    try {
      return alloc();
    } finally {
      throw new IOException();
    }

The allocated resource will be leaked because the method does not
return; it throws an exception.  Previously the resource leak checker
would accept that code because _all_ obligations were cleared at return
statements.  Instead, only normal-return-obligations should be cleared
at return statements, and exceptional-exit-obligations should remain
unsatisfied in case a post-return-finally-block ends up throwing.

This change also affects try-with-resources blocks, where there is an
implicit finally-block that might throw.

In developing this fix I also discovered an interesting false positive,
explained and tested in IfBranch.java.  That false positive is vaguely
related to this fix because it affects the case where there are
multiple return statements appearing in a try-finally block: control
flow merges after the returns, jumping to the common finally-block.
`MustCallConsistencyAnalyzer` does not handle the control flow merge
correctly.