Fix two possible type inference crashes
Fixes #6887.  In a few positions, the type inference code was trying
to obtain the type of `throw E` rather than the type of `E`.

Also fixes another latent bug: the Checker Framework would throw NPE on
the line

    new Issue6887<>(() -> {});

because with no body statements, the various tree visitors in
`CheckedExceptionsUtil` return null.