Fix exponential-time handling of `@This` in CalledMethodsChecker
Fixes #7024.

When run with `-AenableReturnsReceiverForRlc`, the resource leak
checker takes exponential time to check code with long chains of
`@This` methods.

The root cause lies in `CalledMethodsChecker`, which inadvertently
violates this rule from the manual:

> A subclass of `TreeAnnotator` should not, in general, call
> `getAnnotatedType()` more than once. [...] expressions at the base of
> a large binary tree will be visited exponentially many times, leading
> to slowdowns or the appearance of an infinite loop.

The violation is not obvious:

`CalledMethodsAnnotatedTypeFactory.createTreeAnnotator()` returns a
`ListTreeAnnotator` that includes both an `AccumulationTreeAnnotator`
and a `CalledMethodsTreeAnnotator`.  But of those tree annotators call
`getAnnotatedType()` from `visitMethodInvocation()`;  therefore we have
inadvertently built a composite tree annotator that makes two calls to
`getAnnotatedType()` when visiting a method invocation.

The quickest fix is to simply remove one of those invocations.  The one
in `CalledMethodsTreeAnnotator` can easily be avoided by breaking its
inheritance relationship with `AccumulationTreeAnnotator`.  Since
`AccumulationTreeAnnotator` already appears in the list of tree
annotators, there isn't any reason for `CalledMethodsTreeAnnotator` to
extend it.