Fix a false positive in the Resource Leak Checker
Previously, the RLC would report an error on this code:

    try {
      alloc().close();
    } catch (IOException e) {
      throw e;
    }

It would report that the result of `alloc()` was not closed if
`close()` throws IOException, which is clearly incorrect.

That test case is meant to be handled by `visitMethodInvocation()` in
`RLCCalledMethodsTransfer`, which looks at whether the receiver of a
method call has a corresponding temporary variable (which will be true
for expressions like `alloc()` that are not normally tracked by flow
analysis).

However, the `visitMethodInvocation()` override did not consider
exceptional paths.  These were ignored due to an awkward construct in
CalledMethodsTransfer.java: a private `exceptionalStores` field
enabled `accumulate()` to record information along exceptional control
flow edges, but only temporarily while `super.visitMethodInvocation()`
was running.  The overridden `visitMethodInvocation()` in
RLCCalledMethodsTransfer.java makes additional calls to `accumulate()`
after the call to `super` returns, but that work did not properly
update the exceptional stores because after the super method returns
the private field has been unset.

This commit fixes the problem (and makes the code a little less
confusing) by removing the awkward private field.  The behavior of
`accumulate()` now depends on whether accumulate's `result` argument
has exceptional stores, not on the value of a private field.

I don't think the new code is 100% wart-free, but it is an improvement.